+++ packages/nodejs/CHANGELOG.md
@@ -1,18 +1,22 @@
 # Changelog
 
-## Unreleased
+## 6.2.0 2023-2-23
 
 ### Added
 
-- Added a `createConcordiumClient` function to create the GRPC v2 client.
+- Added a `createConcordiumClient` function to create the gRPC v2 client.
 
 ### Changed
 
-- Bumped @concordium/common-sdk to 6.1.0. (adds support for id statements and proofs)
+- Bumped @concordium/common-sdk to 6.3.0. (Adds the gRPC v2 client)
 
 ### Fixed
 
 - The value of amount fields in the GRPCv1 client's invokeContract's events has been changed to bigint (instead of string) as the type specifies.
+
+### Deprecated
+
+- The old gRPC client has been deprecated in favor of the new gRPC v2 client.
 
 ## 6.1.0 2022-11-30